PCto3D/requirements.txt
open3d>=0.18.0
numpy>=1.24.0
scipy>=1.10.0
pillow>=10.0.0
pyyaml>=6.0
tqdm>=4.65.0
scikit-learn>=1.3.0
matplotlib>=3.7.0

